feat: support menu api

+++ migrations/20250508021428_initial.sql
@@ -1,6 +1,6 @@
 -- Add migration script here
 CREATE TABLE IF NOT EXISTS users (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name TEXT NOT NULL UNIQUE,
     email TEXT NOT NULL DEFAULT '',
     password TEXT NOT NULL DEFAULT '',
@@ -10,9 +10,9 @@ CREATE TABLE IF NOT EXISTS users (
     remark TEXT NOT NULL DEFAULT '',
     last_login_time TIMESTAMPTZ NOT NULL,
     is_deleteable BOOLEAN NOT NULL DEFAULT FALSE,
-    organization_id INTEGER NOT NULL DEFAULT 0,
+    organization_id BIGINT NOT NULL DEFAULT 0,
     organization_name TEXT NOT NULL DEFAULT '',
-    role_id INTEGER NOT NULL DEFAULT 0,
+    role_id BIGINT NOT NULL DEFAULT 0,
     role_name TEXT NOT NULL DEFAULT '',
     created_at TIMESTAMPTZ,
     updated_at TIMESTAMPTZ,
@@ -20,10 +20,10 @@ CREATE TABLE IF NOT EXISTS users (
 );
 
 CREATE TABLE IF NOT EXISTS role (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name TEXT NOT NULL UNIQUE,
     description TEXT NOT NULL DEFAULT '',
-    created_by INTEGER NOT NULL DEFAULT 0,
+    created_by BIGINT NOT NULL DEFAULT 0,
     created_by_name TEXT NOT NULL DEFAULT '',
     is_deleteable BOOLEAN NOT NULL DEFAULT FALSE,
     created_at TIMESTAMPTZ,
@@ -32,9 +32,9 @@ CREATE TABLE IF NOT EXISTS role (
 );
 
 CREATE TABLE IF NOT EXISTS menu (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name TEXT NOT NULL UNIQUE,
-    parent_id INTEGER NOT NULL DEFAULT 0,
+    parent_id BIGINT NOT NULL DEFAULT 0,
     parent_name TEXT NOT NULL DEFAULT '',
     order_index INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMPTZ,
@@ -43,10 +43,10 @@ CREATE TABLE IF NOT EXISTS menu (
 );
 
 CREATE TABLE IF NOT EXISTS role_menu (
-    id SERIAL PRIMARY KEY,
-    role_id INTEGER NOT NULL DEFAULT 0,
+    id BIGINT PRIMARY KEY,
+    role_id BIGINT NOT NULL DEFAULT 0,
     role_name TEXT NOT NULL DEFAULT '',
-    menu_id INTEGER NOT NULL DEFAULT 0,
+    menu_id BIGINT NOT NULL DEFAULT 0,
     menu_name TEXT NOT NULL DEFAULT '',
     created_at TIMESTAMPTZ,
     updated_at TIMESTAMPTZ,
@@ -54,9 +54,9 @@ CREATE TABLE IF NOT EXISTS role_menu (
 );
 
 CREATE TABLE IF NOT EXISTS organization (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name TEXT NOT NULL UNIQUE,
-    parent_id INTEGER NOT NULL DEFAULT 0,
+    parent_id BIGINT NOT NULL DEFAULT 0,
     parent_name TEXT NOT NULL DEFAULT '',
     created_at TIMESTAMPTZ,
     updated_at TIMESTAMPTZ,
@@ -64,10 +64,10 @@ CREATE TABLE IF NOT EXISTS organization (
 );
 
 CREATE TABLE IF NOT EXISTS route (
-    id SERIAL PRIMARY KEY,
+    id BIGINT PRIMARY KEY,
     name TEXT NOT NULL UNIQUE,
     method TEXT NOT NULL DEFAULT '',
-    menu_id INTEGER NOT NULL DEFAULT 0,
+    menu_id BIGINT NOT NULL DEFAULT 0,
     menu_name TEXT NOT NULL DEFAULT '',
     created_at TIMESTAMPTZ,
     updated_at TIMESTAMPTZ,
